Fix MiddlewareFunction interface
1. Change client request/response to server request/response, because
the client ones are for when node is acting as a client (making
outgoing requests), whereas here it’s acting as a server. This change
is needed to work with express.

2. Change the `next` argument, because next (in express) is a function
called with 0 or 1 arguments, where that argument can be an any (for
the error). Next is decidedly *not* a function with the
MiddlewareFunction signature.